Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/module/instantiation-error-6-expected.txt

CONSOLE MESSAGE: SyntaxError: Importing binding name 'foo' cannot be resolved due to ambiguous multiple bindings.

FAIL Test that ambiguous star exports lead to an instantiation error and that the correct module is blamed. null is not an object (evaluating 'exn.constructor')
